+++ pac4j-core/src/main/java/org/pac4j/core/engine/CallbackLogic.java
@@ -19,12 +19,10 @@ public interface CallbackLogic {
      * @param config the security configuration
      * @param httpActionAdapter the HTTP action adapter
      * @param defaultUrl the default url
-     * @param saveInSession whether profile should be saved in session
-     * @param multiProfile whether multi profiles are supported
      * @param renewSession whether the session must be renewed
-     * @param client the default client
+     * @param defaultClient the default client
      * @return the resulting action of the callback
      */
     Object perform(WebContext context, Config config, HttpActionAdapter httpActionAdapter,
-                     String defaultUrl, Boolean saveInSession, Boolean multiProfile, Boolean renewSession, String client);
+                     String defaultUrl, Boolean renewSession, String defaultClient);
 }